Added checks for JavaScript enabled or not.
I have also fixed the read more and show less not working properly which is in previous versions, clicking read more would show all the abstract and content in all rows and show less would hide all the abstract and content in all rows. But in this fix, it would show or hide the specific abstract or content based on "Read more" or "Show less" user has clicked on which specific column.